Complete ALTO Entity Property Type Registry
Clearly define the concept Property Type in section 2 and differentiate
it to Property Name.

Regardless the consistency with the Endpoint Property Type Registry. It
is not necessary to enforce the consistency. Endpoint Property Type is a
complex concept including the global property and resource-specific
property. We do not want to introduce such complexity into the unified
properties document.

Signed-off-by: jensenzhang <jingxuan.n.zhang@gmail.com>